\name{9.1.example.data}
\alias{9.1.example.data}
\alias{example.data}
\alias{pems.1}


\docType{data}

\title{
   example data for use with pems.utils
}

\description{
   Example data intended for use with functions in \code{pems.utils}. 
}

\usage{

   pems.1

}


\format{
  \code{pems.1} is a example \code{pems} object.   

}

\details{
  \code{pems.1} is supplied as part of the \code{pems.utils} package. 

}

\note{
  None at present
}

\source{
  Reference in preparation
}

\references{
  None at present
}

\seealso{
  See examples in \code{\link{pems.structure}}. 
}

%\examples{
%#to be confirmed
%}

\keyword{datasets}
